refact: root(Meeting) 중심 설계 적용
